move common token verification to app.main.auth
